+++ apps/client/components/ui/types/button.ts
@@ -1,0 +1,8 @@
+export type Image = {
+    width?: string;
+    margin?: string;
+    imageUrl?: string;
+    objectFit?: string;
+    borderRadius?: string;
+    cursor?: string;
+  };
