+++ apps/client/components/ui/types/commonElements.ts
@@ -1,0 +1,14 @@
+export type Text = {
+    color?: string;
+    width?: string;
+    padding?: string;
+    fontSize?: string;
+    whiteSpace?: string;
+    overflow?: string;
+    textOverflow?: string;
+    cursor?: string;
+    display?: string;
+    marginLeft?: string;
+    fontWeight?: string;
+    margin?: string;
+  };
