+++ apps/client/components/ui/types/testContent.ts
@@ -1,0 +1,31 @@
+export type TitleAndTextProps = {
+  text: {
+    titleText?: string;
+    contentText?: string;
+  };
+};
+
+export type TitleAndTestProps = {
+  style: {
+    titleText?: string;
+    imageUrl?: string;
+    squareText?: string;
+  };
+};
+
+export type TitleAndTestsSmallForSeveralProp = {
+  testData: {
+    testCoverDTOList: {
+      commentCount?: number;
+      id?: string;
+      imageUrl?: string;
+      likeCount?: number;
+      playCount?: number;
+      title?: string;
+      type?: string;
+    }[];
+  };
+  style: {
+    titleText: string;
+  };
+};
