+++ apps/client/components/ui/types/wrap.ts
@@ -1,0 +1,28 @@
+export type WrapMediaqueryProp = {
+  width?: string;
+  margin?: string;
+  flexDirection?: string;
+  justifyContent?: string;
+  alignitems?: string;
+  padding?: string;
+  position?: string;
+  textAlign?: string;
+};
+
+export type Wrap = {
+  width?: string;
+  height?: string;
+  margin?: string;
+  overflow?: string;
+  display?: string;
+  flexWrap?: string;
+  position?: string;
+  right?: string;
+  bottom?: string;
+  padding?: string;
+  flexDirection?: string;
+  justifyContent?: string;
+  alignItems?: string;
+  backgroundColor?: string;
+  borderRadius?: string;
+};
